Add files via upload
Analyzed Criminal Activities in India using Python, Excel and Tableau: I analyzed the data revealing the crimes 
in India against women and Children. The data set reveals the shocking and breathe taking results which raised a 
million of questions for the judicial and crime branches. I used Python Jupyter Notebook for data cleaning, 
manipulation and Visualization. The Python libraries I used in this Project is Pandas, Seaborn, NumPy, summary tool, 
matplotlib for generating graphs and Tableau for visualizations from the data set.